+++ projects/ui-framework/bob-rte/src/rte/rte.abstract.ts
@@ -67,7 +67,6 @@ export abstract class RTEbaseElement extends BaseFormElement
     super(cd);
     this.baseValue = '';
     this.wrapEvent = false;
-    this.ignoreEvents = [InputEventType.onWrite, InputEventType.onFocus];
   }
 
   public tribute: TributeInstance;
