+++ projects/ui-framework/src/public_api.ts
@@ -149,9 +149,8 @@ export { MultiListAndChipsComponent } from './lib/chips/multi-list-and-chips/mul
 
 // FormElements Module
 export { FormElementsModule } from './lib/form-elements/form-elements.module';
-export {
-  FormEvents, InputEventType,
-} from './lib/form-elements/form-elements.enum';
+export * from './lib/form-elements/form-elements.enum';
+export * from './lib/form-elements/form-elements.const';
 export { BaseFormElement } from './lib/form-elements/base-form-element';
 // Textarea Module
 export { TextareaModule } from './lib/form-elements/textarea/textarea.module';
